fix(autocomplete): udpate samples #922
