+++ src/app/data-entries/autocomplete/movie/movie.component.scss
@@ -9,6 +9,7 @@
 
 .igx-date-time > * {
     width: 40%;
+    margin-top: 32px;
 }
 
 .input-group-form {
