+++ src/app/data-entries/data-entries.module.ts
@@ -6,8 +6,12 @@ import { IgxAutocompleteModule, IgxButtonGroupModule, IgxButtonModule, IgxCheckb
     IgxInputGroupModule, IgxNavbarModule, IgxProgressBarModule, IgxRadioModule, IgxRippleModule, IgxSelectModule,
     IgxSwitchModule, IgxTimePickerModule, IgxToggleModule} from "igniteui-angular";
 import { AutocompleteRemote } from "./autocomplete/autocomplete-remote/autocomplete-remote.component";
-import { AutocompleteBasic } from "./autocomplete/autocomplete/autocomplete.component";
-import { MovieComponent } from "./autocomplete/movie/movie.component";
+import { AutocompleteBasic, AutocompletePipeStartsWith } from "./autocomplete/autocomplete/autocomplete.component";
+import {
+    AutocompleteGroupPipeContains,
+    AutocompletePipeContains,
+    MovieComponent
+} from "./autocomplete/movie/movie.component";
 import { ButtonGroupSample1Component } from "./buttonGroup/button-group-sample-1/button-group-sample-1.component";
 import { ButtonGroupSample2Component } from "./buttonGroup/button-group-sample-2/button-group-sample-2.component";
 import { ButtonGroupSample3Component } from "./buttonGroup/button-group-sample-3/button-group-sample-3.component";
@@ -33,7 +37,6 @@ import { InputGroupSample4Component } from "./input-group/input-group-sample-4/i
 import { InputGroupSample5Component } from "./input-group/input-group-sample-5/input-group-sample-5.component";
 import { InputGroupSample6Component } from "./input-group/input-group-sample-6/input-group-sample-6.component";
 import { ReactiveFormsSampleComponent } from "./input-group/reactive-forms/reactive-forms.component";
-import { AutocompleteGenrePipeContains, AutocompletePipeContains, AutocompletePipeStartsWith } from "./pipes/pipes";
 import { RadioGroupSampleComponent } from "./radio/radio-group-sample/radio-group-sample.component";
 import { RadioSample1Component } from "./radio/radio-sample-1/radio-sample-1.component";
 import { RadioSample2Component } from "./radio/radio-sample-2/radio-sample-2.component";
@@ -46,7 +49,7 @@ import { SwitchSample2Component } from "./switch/switch-sample-2/switch-sample-2
         AutocompleteRemote,
         AutocompletePipeContains,
         AutocompletePipeStartsWith,
-        AutocompleteGenrePipeContains,
+        AutocompleteGroupPipeContains,
         ButtonGroupSample1Component,
         ButtonGroupSample2Component,
         ButtonGroupSample3Component,
